todoapp/requirements.txt
asgiref                3.8.1   
certifi                2024.2.2
charset-normalizer     3.3.2   
defusedxml             0.7.1   
Django                 4.2.13  
django-bootstrap-icons 0.8.7   
djangorestframework    3.15.1  
idna                   3.7     
pip                    21.2.4  
requests               2.31.0  
setuptools             58.1.0  
sqlparse               0.5.0   
typing_extensions      4.11.0  
tzdata                 2024.1  
urllib3                2.2.1 